Move the videos list to under /videos/all
