+++ src/routes/videos/+page.ts
@@ -1,15 +1,7 @@
-import { browser } from '$app/environment'
-import { dataStore } from '$lib/data'
+import { redirect } from '@sveltejs/kit'
+import { base } from '$app/paths'
 import type { PageLoad } from './$types'
 
-export const load = (({ params, url }) => {
-	let pageNumber: number = 1
-	if (browser && url.searchParams.get('page')) {
-		pageNumber = parseInt(url.searchParams.get('page') ?? '1') || 1
-	}
-
-	return {
-		pageNumber,
-		videos: dataStore.getVideos(),
-	}
+export const load = (() => {
+	redirect(302, `${base}/videos/shows`)
 }) satisfies PageLoad
